removed WifiDirectManager.java
