--- AndroidApps/android-core/src/main/java/net/discdd/wifidirect/WifiDirectStateListener.java
@@ -1,5 +1,0 @@
-package net.discdd.wifidirect;
-
-public interface WifiDirectStateListener {
-    void onReceiveAction(WifiDirectManager.WifiDirectEvent action);
-}
